Training_center (spring)/src/main/java/com/krylova/service/FacultyService.java
package com.krylova.service;

import com.krylova.entity.Faculty;
import com.krylova.repository.FacultyRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class FacultyService {

    @Autowired
    private FacultyRepository facultyRepository;

    public Faculty create(Faculty faculty) {
        return facultyRepository.save(faculty);
    }

    public Faculty update(Faculty faculty) {
        return facultyRepository.save(faculty);
    }

    public boolean delete(Faculty faculty) {
        facultyRepository.delete(faculty);
        return true;
    }

    public List<Faculty> findAll() {
        return facultyRepository.findAll();
    }

    public Optional<Faculty> findById(Long id) {
        return facultyRepository.findById(id);
    }

    public List<Faculty> findByEmail(String email) {
        return facultyRepository.findByEmail(email);
    }

}
